[CacheBundle] Move to Symfony NS
