+++ src/Symfony/Bundle/CacheBundle/Resources/meta/LICENSE
